creation of my user

+++ mlnba-ss/src/main/java/com/mgu/mlnba/MlnbaApplication.java
@@ -71,25 +71,28 @@ public class MlnbaApplication {
             .subscribe(t -> System.out.println("user: " + t.getUsername()));
     }
     
-//    @EventListener(ApplicationReadyEvent.class)
+    @EventListener(ApplicationReadyEvent.class)
     public void createMarc() {
-        memberRepo.deleteAll().block();
-        roleRepo.deleteAll().block();
         
-        createRoles();
-        
-        Member m = new Member();
-        m.setUsername("marc");
-        m.setLastname("guerrini");
-        m.setFirstname("marc");
-        m.setPassword(PasswordEncoderFactories.createDelegatingPasswordEncoder().encode("password"));
-        
-        roleRepo.findById("role_admin")
-            .subscribe(r -> {
-                m.setRoles(Arrays.asList(r));
-                memberRepo.save(m).block();
-            });
-        
+        Member m2 = memberRepo.findByUsername("marc").block();
+        if(m2==null) {
+            memberRepo.deleteAll().block();
+            roleRepo.deleteAll().block();
+            
+            createRoles();
+            
+            Member m = new Member();
+            m.setUsername("marc");
+            m.setLastname("guerrini");
+            m.setFirstname("marc");
+            m.setPassword(PasswordEncoderFactories.createDelegatingPasswordEncoder().encode("password"));
+            
+            roleRepo.findById("role_admin")
+                .subscribe(r -> {
+                    m.setRoles(Arrays.asList(r));
+                    memberRepo.save(m).block();
+                });
+        }
         System.out.println("Application Ready Event is successfully Started");
         memberRepo.findAll().subscribe(System.out::println);
